# backend/swagger.yaml
openapi: 3.0.3
info:
  title: SaaS Multi-Tenant Backend Platform API
  description: |
    A comprehensive multi-tenant SaaS backend platform built with Flask.

    ## Features
    - Multi-tenant architecture with database isolation
    - JWT-based authentication with refresh tokens
    - File management with MD5 deduplication
    - Role-based access control (Admin, User, Viewer)
    - Document management system
    - Kafka integration (placeholder for Phase 6)

    ## Authentication
    Most endpoints require JWT authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Rate Limiting
    API rate limiting will be implemented in Phase 6.

    ## Error Responses
    All endpoints return consistent error responses with the following structure:
    ```json
    {
      "success": false,
      "message": "Error description",
      "error": "ERROR_CODE"
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4999
    description: Development server
  - url: http://localhost:8080
    description: Docker development server
  - url: https://api.example.com
    description: Production server (Phase 7)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
    Error:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        error:
          type: string
          example: "ERROR_CODE"
...
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "Acme Corp"
        database_name:
          type: string
          example: "tenant_acme_corp_a1b2c3d4"
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        role:
          type: string
          enum: [admin, user, viewer]
          example: "admin"
        joined_at:
          type: string
          format: date-time
...
    SSOConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        tenant_id:
          type: string
          format: uuid
          example: "456e7890-e89b-12d3-a456-426614174000"
        provider:
          type: string
          enum: [azure_ad]
          example: "azure_ad"
        client_id:
          type: string
          example: "your-azure-client-id"
        provider_tenant_id:
          type: string
          example: "your-azure-tenant-id"
        is_enabled:
          type: boolean
          example: true
        config_metadata:
          type: object
          properties:
            auto_provisioning:
              type: object
              properties:
                enabled:
                  type: boolean
                  example: true
                default_role:
                  type: string
                  enum: [admin, user, viewer]
                  example: "viewer"
                sync_attributes_on_login:
                  type: boolean
                  example: true
                allowed_email_domains:
                  type: array
                  items:
                    type: string
                  example: ["@company.com"]
                allowed_azure_groups:
                  type: array
                  items:
                    type: string
                  example: ["All-Employees"]
                group_role_mapping:
                  type: object
                  additionalProperties:
                    type: string
                  example:
                    IT-Admins: "admin"
                    Developers: "user"
                    Readers: "viewer"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /api/tenants:
    get:
      tags:
        - Tenants
      summary: List user's tenants
      description: Returns all tenants the authenticated user is a member of
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Tenants retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Tenants retrieved successfully"
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'

    post:
      tags:
        - Tenants
      summary: Create a new tenant
      description: Creates a new tenant organization. Creator becomes admin.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: "Acme Corp"
      responses:
        '201':
          description: Tenant created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Tenant created successfully"
                  data:
                    $ref: '#/components/schemas/Tenant'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/tenants/{tenant_id}/sso/config:
    get:
      tags:
        - Tenants
        - SSO
      summary: Get SSO configuration
      description: |
        Retrieves the SSO configuration for a tenant.
        Only tenant admins can view the configuration.
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the tenant
      responses:
        '200':
          description: SSO configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "SSO configuration retrieved successfully"
                  data:
                    $ref: '#/components/schemas/SSOConfig'
        '404':
          description: SSO not configured for tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Tenants
        - SSO
      summary: Create SSO configuration
      description: |
        Creates a new SSO configuration for a tenant.
        Only tenant admins can create SSO configuration.
        Note: Client secret is not required for Azure AD public application mode.
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - client_id
                - provider_tenant_id
              properties:
                provider:
                  type: string
                  enum: [azure_ad]
                  default: azure_ad
                  description: SSO provider type
                client_id:
                  type: string
                  description: Azure AD Application (client) ID
                  example: "your-azure-client-id"
                provider_tenant_id:
                  type: string
                  description: Azure AD Directory (tenant) ID
                  example: "your-azure-tenant-id"
                enable:
                  type: boolean
                  default: false
                  description: Enable SSO immediately after creation
                config_metadata:
                  type: object
                  description: Additional configuration options
                  properties:
                    auto_provisioning:
                      type: object
                      properties:
                        enabled:
                          type: boolean
                          default: false
                        default_role:
                          type: string
                          enum: [admin, user, viewer]
                          default: viewer
                        sync_attributes_on_login:
                          type: boolean
                          default: true
                        allowed_email_domains:
                          type: array
                          items:
                            type: string
                          example: ["@company.com"]
                        allowed_azure_groups:
                          type: array
                          items:
                            type: string
                          example: ["All-Employees"]
                        group_role_mapping:
                          type: object
                          additionalProperties:
                            type: string
                          example:
                            IT-Admins: "admin"
      responses:
        '201':
          description: SSO configuration created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "SSO configuration created successfully"
                  data:
                    $ref: '#/components/schemas/SSOConfig'
        '400':
          description: SSO already configured or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Tenants
        - SSO
      summary: Update SSO configuration
      description: |
        Updates the SSO configuration for a tenant.
        Only tenant admins can update SSO configuration.
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                client_id:
                  type: string
                  description: Azure AD Application (client) ID
                provider_tenant_id:
                  type: string
                  description: Azure AD Directory (tenant) ID
                config_metadata:
                  type: object
                  description: Additional configuration options
      responses:
        '200':
          description: SSO configuration updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "SSO configuration updated successfully"
                  data:
                    $ref: '#/components/schemas/SSOConfig'
        '404':
          description: SSO configuration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Tenants
        - SSO
      summary: Delete SSO configuration
      description: |
        Deletes the SSO configuration for a tenant.
        This will disable SSO and remove all stored tokens.
        Only tenant admins can delete SSO configuration.
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the tenant
      responses:
        '200':
          description: SSO configuration deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "SSO configuration deleted successfully"
        '404':
          description: SSO configuration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tenants/{tenant_id}/sso/config/enable:
    post:
      tags:
        - Tenants
        - SSO
      summary: Enable SSO
      description: |
        Enables SSO for a tenant.
        SSO must be configured before it can be enabled.
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the tenant
      responses:
        '200':
          description: SSO enabled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "SSO enabled successfully"
                  data:
                    type: object
                    properties:
                      is_enabled:
                        type: boolean
                        example: true
        '404':
          description: SSO configuration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tenants/{tenant_id}/sso/config/disable:
    post:
      tags:
        - Tenants
        - SSO
      summary: Disable SSO
      description: |
        Disables SSO for a tenant.
        Users can still login with local passwords if available.
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the tenant
      responses:
        '200':
          description: SSO disabled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "SSO disabled successfully"
                  data:
                    type: object
                    properties:
                      is_enabled:
                        type: boolean
                        example: false
        '404':
          description: SSO configuration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tenants/{tenant_id}/sso/config/validate:
    get:
      tags:
        - Tenants
        - SSO
      summary: Validate SSO configuration
      description: |
        Validates the SSO configuration by checking:
        - Azure AD app registration exists
        - Redirect URI is configured correctly
        - Required permissions are granted
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the tenant
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  valid:
                    type: boolean
                    example: true
                  checks:
                    type: object
                    properties:
                      app_registration:
                        type: boolean
                        description: Azure AD app exists
                      redirect_uri:
                        type: boolean
                        description: Redirect URI configured
                      permissions:
                        type: boolean
                        description: Required permissions granted
                      public_client:
                        type: boolean
                        description: Public client flow enabled
                  errors:
                    type: array
                    items:
                      type: string
                    example: []
                  warnings:
                    type: array
                    items:
                      type: string
                    example: ["Admin consent may be required"]
        '404':
          description: SSO configuration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...